// src/Types/Responses.ts
export enum SUCCESS_TYPES {
    CREATED = "Fue creado correctamente",
    DELETED = "Fue eliminado correctamente",
    UPDATED = "Fue actualizado correctamente",
    GETTED = "Correcto"
}

export enum ERROR_TYPES {
    BAD = "Solicitud incorrecta",
    UNAUTHORIZED = "No tienes permiso",
    FORBIDDEN = "Prohibido",
    NOT_FOUND = "Recurso no encontrado",
    IN_USE = "El dato ya esta en uso",
    INTERNAL = "Error interno del servidor"
}